Cambios en iconos y algún estilo En los test y en las clases del Jackson Movido de modelo a eventManager /json Intento de leer eventos pasados

+++ hermesManager/model/src/main/java/es/udc/lbd/hermes/model/events/measurement/Measurement.java
@@ -48,7 +48,7 @@ public class Measurement implements Serializable{
 //      @Enumerated(EnumType.STRING)
         private String tipo;
         
-        private int value;
+        private Double value;
 
         @ManyToOne(fetch = FetchType.EAGER)
     	@JoinColumn(name = "idUsuario")
@@ -97,11 +97,11 @@ public class Measurement implements Serializable{
 			this.tipo = tipo;
 		}
 
-		public int getValue() {
+		public Double getValue() {
 			return value;
 		}
 
-		public void setValue(int value) {
+		public void setValue(Double value) {
 			this.value = value;
 		}
